Refactor: Remove LauncherCoordinator and ManualSpinController
- Move artifact color tracking to IntakeSubsystem (add getArtifactCount())
- Update IntakeUntilFullCommand for autonomous intake automation
- Delete ManualSpinController interface (no longer needed)
- Delete LauncherCoordinator subsystem (responsibilities distributed)
- Update Robot.java to wire lighting directly to intake
- Update LauncherCommands factory to remove coordinator/manual spin params
- Update FireAllAtRangeCommand, FireAllAtDistanceCommand, FireAllCommand
- Update LaunchObeliskPatternCommand to use IntakeSubsystem for color tracking

Still TODO (compilation fixes needed):
- FireAllAtAutoRangeCommand.java
- FireLaneCommand.java
- SpinHoldCommand.java
- SpinLaneCommand.java
- Telemetry data classes (TelemetryService, IntakeTelemetryData, etc.)
- Update bindings/opmodes

Architecture improvements:
- No auto-spin management (explicit command control)
- IntakeUntilFullCommand for autonomous (no periodic automation)
- Simpler dependency graph
- Lighting wired directly at init
