+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/commands/IntakeCommands/IntakeUntilFullCommand.java
@@ -5,6 +5,12 @@ import org.firstinspires.ftc.teamcode.subsystems.IntakeSubsystem;
 
 /**
  * Runs the intake until all lanes are filled (or until a timeout expires).
+ * Automatically manages intake mode:
+ * - ACTIVE_FORWARD when fewer than 3 artifacts detected
+ * - PASSIVE_REVERSE when 3 artifacts detected (full)
+ *
+ * Use this command in autonomous for intelligent intake automation.
+ * In teleop, operators should have direct manual control instead.
  */
 public class IntakeUntilFullCommand extends IntakeCommand {
 
@@ -18,16 +24,26 @@ public class IntakeUntilFullCommand extends IntakeCommand {
 
     @Override
     public void start() {
-        getIntake().setMode(IntakeSubsystem.IntakeMode.ACTIVE_FORWARD);
         timer.reset();
     }
 
     @Override
-    public boolean isDone() {
-        // For now, timeout only
-        if (timer.seconds() >= timeoutSeconds) return true;
+    public void execute() {
+        // Auto-manage intake mode based on artifact count
+        int count = getIntake().getArtifactCount();
+        if (count >= 3) {
+            getIntake().setMode(IntakeSubsystem.IntakeMode.PASSIVE_REVERSE);
+        } else {
+            getIntake().setMode(IntakeSubsystem.IntakeMode.ACTIVE_FORWARD);
+        }
+    }
 
-        return getIntake().isFull();
+    @Override
+    public boolean isDone() {
+        if (timer.seconds() >= timeoutSeconds) {
+            return true;
+        }
+        return getIntake().getArtifactCount() >= 3;
     }
 
     @Override
